Extract clipboard and touch handlers from use-terminal.ts
Move clipboard (copy/paste/image-sync) and touch/mouse event handlers
from the monolithic xterm init effect into terminal-event-handlers.ts.
Each setup function returns a cleanup callback, keeping the init effect
focused on terminal creation and lifecycle. Reduces use-terminal.ts
from 1225 to 1071 lines with no behavior change.
